requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.3
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
